Align footer counters with spec uint64 typing
Spec §5.5 declares total_chunk_count and total_plaintext_bytes as
uint64 BE, but the reader and footer model previously parsed them as
signed long, narrowing the valid range. Switch PqfFooter,
PqfFileReader, and the decryptor comparison sites to ulong, and update
PqfFileWriterTests to match. Also drop the duplicate
DeterministicCborValidator.ParseStrict call in PqfFileReader.

+++ src/PostQuantum.FileFormat/File/StreamingModeDecryptor.cs
@@ -63,7 +63,7 @@ internal sealed class StreamingModeDecryptor
                     SecureZero.Clear(plaintextBuffer);
                 }
 
-                if (emitted != reader.ReportedPlaintextBytes)
+                if ((ulong)emitted != reader.ReportedPlaintextBytes)
                 {
                     return new PqfDecryptResult(false, PqfRefusalReason.FooterPlaintextBytesMismatch, emitted, true);
                 }
